requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
celery==5.4.0
redis==5.0.4
pydantic==2.9.2
fsspec==2024.2.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
streamlit==1.40.0
requests==2.32.3
flower==2.0.1
moviepy==1.0.3
numpy<2.0.0
pytest==8.3.3
gTTS==2.5.1
edge-tts==6.1.12
pillow==10.4.0
pillow-avif-plugin==1.4.3
jinja2==3.1.4
# Color Book Module
google-generativeai
google-cloud-aiplatform
reportlab
pandas
openpyxl
opencv-python
jsonschema

